src/requirements.txt
click
python-qpid-proton
pyngus
python-json-logger
PyYAML
paho-mqtt
